requirements.txt
cwltool==1.0.20180116213856
cwl-runner==1.0
Flask==0.12.2
Flask-Cors==3.0.3
aap-client-python==0.1.1
cryptography==2.1.4
PyYAML==3.12
future==0.16.0
